Require initial zoom/center when constructing Map
Previously, the Map class allowed you to give it a center/zoom lazily.
That introduced a lot of complexity in the library because null
checks/assertions had to be performed anywhere the map view state
was accessed. And for what? Why would you create a map on the page
before you actually know which part of the world you want it to show?

Additionally, some code, such as that for adding layers, listened for a
'load' event from the map. Because the event could come asynchronously,
checks had to be done to make sure the layer was still registered by the
time the map was loaded. Now the 'load' event no longer exists.