Switch to array for Canvas draw order
Previously, the Canvas class used a linked-list to keep track of the
order in which vector layers were to be drawn, with the last layers to
be drawn showing up in front (on top of everything else). This was
not great because:

- Iteration speed for linked-lists is pretty terrible compared to arrays
due to the way modern CPU caches work, and Canvas must iterate
over all of the layers in order every time it renders. It will render 60
times per second for smooth animations, which is WAY more often
than the average application will be rearranging the order in which
layers are drawn--rearranging the order would be faster with a
linked-list if there are many layers.
- Because many of the Canvas methods needed to quickly access
the linked-list node for a particular layer, Canvas was adding a
property (_order) to each Path instance and then deleting the
property whenever a Path was removed from the Map. Many
modern JS engine optimizations depend on object properties NOT
changing like this because they want to associate each object
with a static-ish type for fast property lookups. Not to mention,
this Canvas-specific property cluttered the renderer-agnostic
Path code.
- The linked-list operations tended to be more verbose code-wise
than array operations, bloating the final JS bundle-size of the library.