Improve Canvas/SVG performance
Moved the map/registry of paths from Renderer to SVG because Canvas can
just use the array of paths it already must maintain for draw order.
That means better iteration performance and no more unnecessary map
adds/deletes. Additionally, SVG now uses an ES6 Set instance for
registering paths so it does not need to use the Util.stamp() function,
not to mention we can easily iterate over a Set directly without
allocating an array (previously used Object.values()) or resorting to a
for-in loop.